% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics.R
\name{remove}
\alias{remove}
\title{Remove dependencies}
\usage{
remove(x, dep)
}
\arguments{
\item{x}{A dependency_manager.}

\item{dep}{A dependency key, html_dependency, htmlwidget, or list.}
}
\value{
An updated \code{dependency_manager} with the specified dependencies
removed.
}
\description{
Removes one or more previously registered dependencies and their associated
asset ids from the manager. Does not delete copied files from disk.
}
\seealso{
\code{\link[=has]{has()}}, \code{\link[=insert]{insert()}}, \code{\link[=DependencyManager]{DependencyManager()}}
}
